+++ testhelper/docker/resource/pulsar/pulsar_test.go
@@ -1,4 +1,4 @@
-package resource_test
+package pulsar
 
 import (
 	"net/http"
@@ -8,14 +8,15 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/rudderlabs/rudder-go-kit/httputil"
-	"github.com/rudderlabs/rudder-go-kit/testhelper/docker/resource"
 )
 
 func TestPulsar(t *testing.T) {
 	pool, err := dockertest.NewPool("")
 	require.NoError(t, err)
-	pulsarContainer, err := resource.SetupPulsar(pool, t)
+
+	pulsarContainer, err := SetupPulsar(pool, t)
 	require.NoError(t, err)
+
 	res, err := http.Head(pulsarContainer.AdminURL + "/admin/v2/namespaces/public/default")
 	defer func() { httputil.CloseResponse(res) }()
 	require.NoError(t, err)
